Workarround to keep camera streaming going
The backport of AtomISP on Intel Aero have some problems and it was
crashing the streaming and cause the other events on the same poll
to also block.

So here moving the camera to a separted thread and when the streaming
fails it is reconfiguring the V4L2 that causes the camera streaming
work again.

Also some non-related changed done here, sorry about not spliting
in several patches.